refactor: simplify gardenRoutes, plantRoutes, and seedRoutes by modularizing controllers
