+++ src/routes/seedRoutes.ts
@@ -1,7 +1,6 @@
-import prisma from '@/config/db';
+import { addSeeds, getSeedCount, useSeeds } from '@/controllers/item/seedController';
 import { clientAuth } from '@/middlewares/authMiddleware';
-import { AuthRequest } from '@/types/auth';
-import express, { Response } from 'express';
+import express from 'express';
 
 const router = express.Router();
 
@@ -9,76 +8,12 @@ const router = express.Router();
 router.use(clientAuth);
 
 // Get user's seed count
-router.get('/', async (req: AuthRequest, res: Response) => {
-  try {
-    const seed = await prisma.seed.findUnique({
-      where: { userId: req.user!.id }
-    });
-    
-    res.json(seed || { userId: req.user!.id, count: 0 });
-  } catch (error) {
-    res.status(500).json({ message: 'Error fetching seed count' });
-  }
-});
+router.get('/', getSeedCount);
 
 // Add seeds to user
-router.post('/add', async (req: AuthRequest, res: Response) => {
-  try {
-    const { count } = req.body;
-    
-    if (!count || count <= 0) {
-      return res.status(400).json({ message: 'Valid seed count is required' });
-    }
-    
-    const seed = await prisma.seed.upsert({
-      where: { userId: req.user!.id },
-      update: {
-        count: {
-          increment: count
-        }
-      },
-      create: {
-        userId: req.user!.id,
-        count
-      }
-    });
-    
-    res.json(seed);
-  } catch (error) {
-    res.status(500).json({ message: 'Error adding seeds' });
-  }
-});
+router.post('/add', addSeeds);
 
 // Use seeds
-router.post('/use', async (req: AuthRequest, res: Response) => {
-  try {
-    const { count } = req.body;
-    
-    if (!count || count <= 0) {
-      return res.status(400).json({ message: 'Valid seed count is required' });
-    }
-    
-    const seed = await prisma.seed.findUnique({
-      where: { userId: req.user!.id }
-    });
-    
-    if (!seed || seed.count < count) {
-      return res.status(400).json({ message: 'Not enough seeds' });
-    }
-    
-    const updatedSeed = await prisma.seed.update({
-      where: { userId: req.user!.id },
-      data: {
-        count: {
-          decrement: count
-        }
-      }
-    });
-    
-    res.json(updatedSeed);
-  } catch (error) {
-    res.status(500).json({ message: 'Error using seeds' });
-  }
-});
+router.post('/use', useSeeds);
 
 export default router;
